+++ tests/app/rdev/applications/environments/hosts/HostRegexTest.php
@@ -4,7 +4,7 @@
  *
  * Tests the host regex
  */
-namespace RDev\Applications\Environments;
+namespace RDev\Applications\Environments\Hosts;
 
 class HostRegexTest extends \PHPUnit_Framework_TestCase
 {